feat: add userGroup route

+++ routes/userGroup.js
@@ -21,8 +21,28 @@ router.get('/list', async (ctx) => {
   }
 });
 
+async function hasAccess({ ctx, id, user_id }) {
+  const res = await ctx.codeRepo.getUserOfGroup({ id, user_id });
+
+  if (res.access_level < 50) {
+    return false;
+  }
+
+  // 不能操作自己
+  if (ctx.gitUid === user_id) {
+    return false;
+  }
+
+  return true;
+}
+
 router.post('/delete', async (ctx) => {
   const { gitUid, groupId } = ctx.request.body;
+
+  if (!(await hasAccess({ ctx, id: groupId, user_id: gitUid }))) {
+    ctx.body = { message: '没有操作权限' };
+    return;
+  }
 
   await ctx.codeRepo.removeUserFromGroup({
     group_id: groupId,
@@ -42,6 +62,11 @@ router.post('/save', async (ctx) => {
       where: { uid },
     });
 
+    if (!(await hasAccess({ ctx, id: groupId, user_id: res.gitUid }))) {
+      ctx.body = { message: '没有操作权限' };
+      return;
+    }
+
     await ctx.codeRepo.addUserIntoGroup({
       group_id: groupId,
       user_id: res.gitUid,
@@ -49,6 +74,11 @@ router.post('/save', async (ctx) => {
     });
     ctx.body = { message: '操作成功' };
   } else if (gitUid) {
+    if (!(await hasAccess({ ctx, id: groupId, user_id: gitUid }))) {
+      ctx.body = { message: '没有操作权限' };
+      return;
+    }
+
     // 编辑
     await ctx.codeRepo.editUserInGroup({
       group_id: groupId,
